sync: remove 1 thread limitation on local syncs
This limitation was put in place from the belief that most of the time
would be spent on disk I/O on local syncs anyways.

With growing adoption of fast SSDs, this is no longer true and the
bottleneck moved back to the CPU-bound Git's object unpacking process.

By removing this limitation and allowing multiple Git processes to work
on local syncs, the total elapsed time saw a reduction by 43% and an
almost linear scaling on CPU utilization on a Samsung 960 EVO NVMe SSD
with -j16 on the AOSP's android-11.0.0_r1 tag.

Command: "\time -v repo sync -l -j16"

Before this patch:
	Percent of CPU this job got: 81%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 5:27.95

After this patch:
	Percent of CPU this job got: 1225%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 3:06.18

Change-Id: Iece7010c9516de4c72416ae70d5a405bcafca4d8
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>