HEAD pointing to non-existing ref with worktrees.
When checkout is done with Git worktrees then the HEAD in the bare-git
repositories point to 'refs/heads/master'. This master branch does not
locally exist and is not automatically created.
When a user now creates a branch in any git repository named 'master'
then it is no longer possible to get rid of this branch, neither is it
possible to switch to another branch and switch back to this master
branch. Git concludes the 'master' branch is already checked out
(in the bare Git) and that results in a lockdown of this master branch.

To repoduce this issue, run these commands in a repo tree checked
out with --worktree:
- git checkout master # assuming the remote repo has a master branch,
                      # a local tracking branch master is created here
- git checkout -b temp
- git checkout master # This one now fails
- git checkout -d master # fails too
The failure is caused by Git assuming the master branch is checked out
by the bare git repository since HEAD is pointing towards it.

Signed-off-by: Remy Bohmer <oss@bohmer.net>
Change-Id: I15c96604363c41f0d01c42f533174393097daeb5