Allow path="." when parsing <project>s
The OpenBMC project includes a main top-level project and many optional
sub-projects as directories in the top level. These optional
sub-projects typically exist in different repos, so defining them as
children in the manifest gets messy.

An example project layout for OpenBMC looks like:

<manifest>
  ...
  <project name="openbmc" path="." revision="stable"/>
  <project name="meta-company-internal-patches"/>
  <project name="meta-company-private"/>
</manifest>

This results in a file tree that looks like the openbmc repo with a few
extra directories/repos dropped in the top level.

The recommendation to use copyfile or linkfile are not practical for
OpenBMC since the manifest would have to be manually updated (fragile,
prone to human error, and high probability of merge conflicts between
for otherwise unrelated commits).

Change-Id: I4b4b96056669fd53d666ba9b3acef9605c6632c3