tests: Make the tests pass for Python < 3.8
Before Python 3.8, xml.dom.minidom sorted the attributes of an element
when writing it to a file, while later versions output the attributes in
the order they were created. Adapt the generation of remote and project
elements so that the order in which the attributes are created matches
the sorted order (at least for attributes that are part of any tests).

Bug: https://crbug.com/gerrit/14382
Change-Id: Ie2597727afcc48f9063a7261ad970e8a549f0587