Update linkfile if manifest has been updated
Currently, linkfiles which marked by "<linkfile/>" in
manifest will be created by first exec 'repo sync'.
But if some "<linkfile/>" are removed in manifest, then
'repo sync', these removed "<linkfile/>" dest can not
be removed in the sourcecode workspace.

This patch is intent to fix this issue, by save a
'linkfile.txt' in .repo and then compared with new
linkfiles when next sync. If any "<linkfile/>" was
removed, the dest path will be removed in sourcecode
at the same time.

Bug: https://crbug.com/gerrit/11008
Change-Id: I6b7b41e94df0f9e6e52801ec755951a4c572d05d
Signed-off-by: jiajia tang <tangjiajia@xiaomi.com>