Perform a manifest reload if needed.
This fixes an issue that repo sync might fail when superproject is used
with clone bundles that do provided the object, but not the expected
refs.  Without this change, the later _Checkout() call would be
expecting the manifest revision for a project, which will fail because
during Sync_NetworkHalf we only checked whether the SHA1 was present.

Bug: 191716848
Change-Id: I0e9260a49288a86df0eb85e3fd1689eeaa310fbb