fix: Handle tag ref in "upstream" field (revision locked manifests)
repo sync only handles a git tag properly when it is in the "revision" field. However, "revision locked manifests" (`repo manifest --revision-as-HEAD`) specifies the tag in the "upstream" field. The issue is that this tag is not fetched. Only the commit that the tag points to is fetched. This cases issues as self._CheckForImmutableRevision() runs and comes to the conclusion that the tag was changed while in fact, it was just not fetched. This causes a full sync.

https://gerrit.googlesource.com/git-repo/+/HEAD/docs/manifest-format.md#Element-project
> Attribute upstream: Name of the Git ref in which a sha1 can be found. Used when syncing a revision locked manifest in -c mode to avoid having to sync the entire ref space. Project elements not setting their own upstream will inherit this value.

Change-Id: I0507d3a5f30aee8920a9f820bafedb48dd5db554